replaced mmul by inner-product, added 8 delimiter
The clojure.core.matrix/mmul matrix multiplication
operator has bug: it converts all numbers to doubles.
I need Ratios to be preserved.
But clojure.core/matrix/inner-product has an
extended sense that allows it to implement matrix
multiplicationm, and it doesn't have the bug.
So I'm using that.  Per a suggestion by mikera
in an issue.

Also, I discovered that the last circuit encoding
has to be terminated by 8, which is apparently not
only needed at the beginning of an encoding.

And now it works!