Initial Haskell commit equiv M-L and Martingale randomness
This is code to help verify/understand part of the second half of Downey
and Hirschfeldt's theorem 6.3.4, their version of Schnorr's theorem that
Martin-Löf randomness is equivalent to randomness defined by computably
enumerable martingales.  In particular, a set of infinite sequences is
Martin-Löf iff it contains no sequences such thagt a c.e. martingale
would produce infinite profit betting on subsequent digits of such a
sequence.

(I could have written it in a somewhat similar manner in Clojure, but
I wanted to learn Haskell, and this repo is my collection of algorithmic
randomness-learning/testing/exploring code, so I put it here.)