Cleaned up debugging code after fix of so-called addPayouts bug
The bug wasn't in addPayouts.  It was in truncateTree, in which
I accidentally was always recursing on the left and never on the
right.  I noticed the but in github and fixed it there.  That was
merged in commit 82f743f, but I never ran the code until today,
when I discovered that the bug "in addPayouts" was suddently gone!

I thought the problem had to do with defining an infinite tree
cyclicly (not merely recursively), because it only occurred when
I applied truncateTree *after* running addPayouts or some minimal
analogue of it.  truncateTree makes new nodes, and I thought that
was where the solution was.  I thought that Haskell had a subtle
problem with cyclic trees, and maybe they were just asking for
trouble.

But the actual problem was that when I ran the buggy truncateTree after
running addPayouts, truncateTree duplicated left branches and overwrote
right branches. I never noticed this when I ran truncateTree *before*
applying addPayouts, because in those cases, truncateTree just overwrote
zeros with zeros.